Añadida lógica puntuación

+++ src/main/java/Entities/Jugador.java
@@ -34,10 +34,10 @@ public class Jugador implements Serializable {
     private String nickJugador;
 
     @Column(name = "maxPuntuacionPartida")
-    private Integer maxPuntuacionPartida;
+    private Float maxPuntuacionPartida;
 
     @Column(name = "puntuacionTotal")
-    private Integer puntuacionTotal;
+    private Float puntuacionTotal;
 
     @OneToMany(mappedBy = "jugador")
     private List<Partida> partidaList;
@@ -66,19 +66,19 @@ public class Jugador implements Serializable {
         this.nickJugador = nickJugador;
     }
 
-    public Integer getMaxPuntuacionPartida() {
+    public Float getMaxPuntuacionPartida() {
         return maxPuntuacionPartida;
     }
 
-    public void setMaxPuntuacionPartida(Integer maxPuntuacionPartida) {
+    public void setMaxPuntuacionPartida(Float maxPuntuacionPartida) {
         this.maxPuntuacionPartida = maxPuntuacionPartida;
     }
 
-    public Integer getPuntuacionTotal() {
+    public Float getPuntuacionTotal() {
         return puntuacionTotal;
     }
 
-    public void setPuntuacionTotal(Integer puntuacionTotal) {
+    public void setPuntuacionTotal(Float puntuacionTotal) {
         this.puntuacionTotal = puntuacionTotal;
     }
